ParticleSystem, fire, TextureStyle

+++ src/EntityFactory.hpp
@@ -16,7 +16,7 @@ class EntityFactory {
 public:
 	EntityFactory() = delete;
 	static Entity createPlayer(vec position);
-	static Entity createCamera(vec position, uint8_t zoom);
+	static Entity createCamera(vec position, float zoom);
 	static Entity createAnimal(AnimalId::value animalId, vec position);
 	static Entity createResource(ResourceId::value resourceId, pair position);
